+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -13,6 +13,8 @@
 * Removed `Claims::getByRank` and `Claims::getByRanks` (you can use `StatementList::getWithRank` instead)
 * Removed `Claims::getMainSnaks` (you can use `StatementList::getMainSnaks` instead)
 * Removed `Claims::getClaimsForProperty` (you can use `StatementList::getWithPropertyId` instead)
+* Removed `Claims::getHashes`
+* Removed `Claims::getGuids`
 
 ## Version 2.6.0 (2015-03-08)